Fix type errors and add mypy + format gates to CI
- Fix real bug in models.py:display_name (BasicIndividual/BasicOrganization
  variable reuse masked type error)
- Narrow BasicIndividual | BasicOrganization with isinstance in routes/api.py
- Add typing.cast for Any returns from httpx.json() and supabase result.data
- Rename local NPIResult | None fetch results to satisfy mypy flow analysis
  in cli.py and routes/providers.py
- Add types-passlib to typecheck job
- Disable prop-decorator check (Pydantic v2 @computed_field false positive)
- Add mypy + ruff format --check jobs to ci.yml

+++ src/docstats/client.py
@@ -252,4 +252,4 @@ class NPPESClient:
                     return max(float(retry_after), 0.5)
                 except (ValueError, TypeError):
                     pass
-        return RETRY_BACKOFF_BASE * (2**attempt)
+        return float(RETRY_BACKOFF_BASE * (2**attempt))
